temporary test on on location change
