+++ app/src/main/res/values/called_integers.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <integer name="defaut_int">1</integer>
+
+</resources>
